Add propshaft support
