+++ app/assets/javascripts/spina/application.js
