+++ app/assets/javascripts/spina/controllers/confetti_controller.js
