+++ app/assets/javascripts/spina/controllers/form_controller.js
@@ -1,6 +1,5 @@
 import { Controller } from "@hotwired/stimulus"
 import debounce from "libraries/debounce"
-import formRequestSubmitPolyfill from "libraries/form-request-submit-polyfill"
 
 export default class extends Controller {
