+++ app/assets/javascripts/spina/controllers/hotkeys_controller.js
