+++ app/assets/javascripts/spina/controllers/sortable_controller.js
@@ -1,6 +1,5 @@
 import { Controller } from "@hotwired/stimulus"
 import Sortable from "libraries/sortablejs"
-import formRequestSubmitPolyfill from "libraries/form-request-submit-polyfill"
 
 export default class extends Controller {
   static get targets() {
