+++ app/assets/javascripts/spina/libraries/canvas-confetti.js
@@ -1,1 +1,2 @@
-//= require ./canvas-confetti@1.3.2.js
+import confetti from "libraries/canvas-confetti@1.3.2"
+export default confetti
