--- app/assets/javascripts/spina/libraries/form-request-submit-polyfill.js
@@ -1,1 +1,0 @@
-//= require ./form-request-submit-polyfill@2.0.0.js
