+++ app/assets/javascripts/spina/libraries/sortablejs.js
@@ -1,1 +1,2 @@
-//= require ./sortablejs@1.13.0.js
+import Sortable from "libraries/sortablejs@1.13.0"
+export default Sortable
