+++ lib/spina/railtie.rb
@@ -1,7 +1,10 @@
 module Spina
   class Railtie < Rails::Railtie
     initializer "spina.assets.precompile" do |app|
-      app.config.assets.precompile += %w[spina/manifest]
+      if defined?(Sprockets)
+        # Sprockets configuration
+        app.config.assets.precompile += %w[spina/manifest]
+      end
     end
 
     initializer "spina.theme_reloader" do |app|
